Move CLI into submodule
- add the __main__ so you can still do python -m pykeen.cli
- reorganize CLI utilities
- reorganize prompt_config into its own module
