+++ .coveragerc
@@ -5,6 +5,7 @@ omit =
     tests/*
     docs/*
     scripts/*
+    src/pykeen/cli/*
 
 [paths]
 source =
