+++ src/pykeen/cli/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+"""PyKEEN's command line interface."""
+
+from .cli import main
