+++ src/pykeen/cli/__main__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+"""PyKEEN's command line interface."""
+
+from .cli_main import main
+
+if __name__ == '__main__':
+    main()
