+++ src/pykeen/cli/utils/__init__.py
@@ -2,34 +2,35 @@
 
 """Utilities for the command line interface."""
 
-from pykeen.utilities.cli_utils.conv_e_cli import (
+from .conv_e_cli import (
     configure_conv_e_hpo_pipeline, configure_conv_e_training_pipeline,  # noqa: 401; noqa: 401
 )
-from pykeen.utilities.cli_utils.distmult_cli import (
+from .distmult_cli import (
     configure_distmult_hpo_pipeline, configure_distmult_training_pipeline,  # noqa: 401; noqa: 401
 )
-from pykeen.utilities.cli_utils.ermlp_cli import (
+from .ermlp_cli import (
     configure_ermlp_hpo_pipeline, configure_ermlp_training_pipeline,  # noqa: 401; noqa: 401
 )
-from pykeen.utilities.cli_utils.rescal_cli import (
+from .rescal_cli import (
     configure_rescal_hpo_pipeline, configure_rescal_training_pipeline,  # noqa: 401; noqa: 401
 )
-from pykeen.utilities.cli_utils.structured_embedding_cli import (
+from .structured_embedding_cli import (
     configure_se_hpo_pipeline, configure_se_training_pipeline,  # noqa: 401; noqa: 401
 )
-from pykeen.utilities.cli_utils.trans_d_cli import (
+from .trans_d_cli import (
     configure_trans_d_hpo_pipeline, configure_trans_d_training_pipeline,  # noqa: 401; noqa: 401
 )
-from pykeen.utilities.cli_utils.trans_e_cli import (
+from .trans_e_cli import (
     configure_trans_e_hpo_pipeline, configure_trans_e_training_pipeline,  # noqa: 401; noqa: 401
 )
-from pykeen.utilities.cli_utils.trans_h_cli import (
+from .trans_h_cli import (
     configure_trans_h_hpo_pipeline, configure_trans_h_training_pipeline,  # noqa: 401; noqa: 401
 )
-from pykeen.utilities.cli_utils.trans_r_cli import (
+from .trans_r_cli import (
     configure_trans_r_hpo_pipeline,
     configure_trans_r_training_pipeline,  # noqa: 401; noqa: 401
 )
-from pykeen.utilities.cli_utils.unstructured_model_cli import (
+from .unstructured_model_cli import (
     configure_um_hpo_pipeline, configure_um_training_pipeline,  # noqa: 401; noqa: 401
 )
+
