+++ src/pykeen/cli/utils/cli_query_helper.py
