+++ src/pykeen/cli/utils/conv_e_cli.py
