+++ src/pykeen/cli/utils/distmult_cli.py
