+++ src/pykeen/cli/utils/rescal_cli.py
