+++ src/pykeen/cli/utils/structured_embedding_cli.py
