+++ src/pykeen/cli/utils/trans_d_cli.py
