+++ src/pykeen/cli/utils/trans_e_cli.py
