+++ src/pykeen/cli/utils/trans_h_cli.py
