+++ src/pykeen/cli/utils/trans_r_cli.py
