+++ src/pykeen/cli/utils/unstructured_model_cli.py
