+++ src/pykeen/utilities/summarize.py
